functions/file_processor/requirements.txt
functions-framework
google-cloud-aiplatform
vertexai
google-api-python-client
google-auth-oauthlib
python-docx
PyMuPDF